Rename `imagenet_experiment.py` to `handlers.py`
