+++ nupic/research/frameworks/vernon/__init__.py
@@ -19,4 +19,4 @@
 #
 
 from .common_experiments import *
-from .imagenet_experiment import *
+from .handlers import *
