+++ nupic/research/frameworks/vernon/handlers.py
